+++ spec/directions/bake_chapter_references/main_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe Kitchen::Directions::BakeChapterReferences do
+  it 'calls v1' do
+    expect_any_instance_of(Kitchen::Directions::BakeChapterReferences::V1).to receive(:bake)
+      .with(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', klass: 'klass')
+    described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', klass: 'klass')
+  end
+end